fix(atoms): fix scrolling for consecutive modals
When a modal is closed it removes the class "modal-open" from the body after 500ms.
If another modal is opened during this time, it will set the class "modal-open" but it will be deleted because of the timeout in the first modal.
A MutationObserver was added to fix this issue.